# RAM-backed small-file cache with auto-flush; mount without PUT-Range; cache_threshold option; robust unmount