--- a/osdep_linux.go
+++ b/osdep_linux.go
@@ -1,10 +1,9 @@
 package main
 
 import (
-        "syscall"
+	"syscall"
 )
 
 func Dup2(oldfd int, newfd int) (err error) {
 	return syscall.Dup3(oldfd, newfd, 0)
 }
-